move function from shared pyupgrade/_token_helpers to pyupgrade/_plugins/type_bases_object as it is not shared
